requirements.txt
streamlit==1.35.0
google-generativeai==0.4.1
grpcio==1.62.2
python-dotenv==1.0.1
